agents/site-crawler/requirements.txt
anthropic>=0.39.0
playwright>=1.48.0
supabase>=2.10.0
python-dotenv>=1.0.0
fastapi[standard]>=0.115.0
